# Avoid invoking `run.ps1` and port helper module to C# to improve throughput performance