--- a/src/Microsoft.Azure.Functions.PowerShellWorker.csproj
+++ b/src/Microsoft.Azure.Functions.PowerShellWorker.csproj
@@ -22,7 +22,6 @@ Licensed under the MIT license. See LICENSE file in the project root for full li
   <ItemGroup>
     <PackageReference Include="Grpc.Core" Version="1.20.1" />
     <PackageReference Include="Microsoft.PowerShell.SDK" Version="6.2.0" />
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
     <PackageReference Include="CommandLineParser" Version="2.3.0" />
     <PackageReference Include="Google.Protobuf" Version="3.7.0" />
   </ItemGroup>