--- a/src/Public/Commands/TracePipelineObjectCommand.cs
+++ b/src/Public/Commands/TracePipelineObjectCommand.cs
@@ -28,7 +28,7 @@ namespace Microsoft.Azure.Functions.PowerShellWorker.Commands
         /// <summary>
         /// The object from pipeline.
         /// </summary>
-        [Parameter(Mandatory = true, ValueFromPipeline = true)]
+        [Parameter(Mandatory = true, Position = 0, ValueFromPipeline = true)]
         public object InputObject { get; set; }
 
         private static PowerShell s_pwsh;