… docs/cmdlets/Get-OutputBinding.md
… docs/cmdlets/Push-OutputBinding.md
… docs/cmdlets/Trace-PipelineObject.md
… src/Modules/Microsoft.Azure.Functions.PowerShellWorker/Microsoft.Azure.Functions.PowerShellWorker.psm1
… src/Modules/Microsoft.Azure.Functions.PowerShellWorker/PowerShellWorker.Resource.psd1
… src/Public/FunctionMetadata.cs
… src/RequestProcessor.cs
… src/Utility/Utils.cs
… src/resources/PowerShellWorkerStrings.resx
… test/Unit/Function/FunctionLoaderTests.cs
… test/Unit/Function/TestScripts/FuncWithRequires.ps1
… test/Unit/Modules/HelperModuleTests.cs
… test/Unit/Modules/Microsoft.Azure.Functions.PowerShellWorker.Tests.ps1
… test/Unit/PowerShell/PowerShellManagerTests.cs
… test/Unit/PowerShell/TestScripts/testBasicFunctionSpecialVariables.ps1
… test/Unit/PowerShell/TestScripts/testBasicFunctionWithRequires.ps1
… test/Unit/PowerShell/TestScripts/testBasicFunctionWithTriggerMetadata.ps1
… tools/helper.psm1