Summary
-----------
Fix #127
Note that, the NO.1 change below is a big design change, and I would like your help to think about all scenarios that might not work with this change.
For example, @TylerLeonhardt had the great concern whether this will continue to work with local debugging. I tested it and it still works.

Major changes
----------------

1. Deploy the Azure Function scripts (e.g. `run.ps1`) to `Runspace` as actual PowerShell functions, so as to avoid hitting the file system for every invocation when using `AddCommand(<path-to-ps1-file>).Invoke()`.
    - In the sandbox environment, the FunctionApp files are on the user's storage account, and thus file system access is expensive. Invoking an in-memory PS function is 8x faster than invoking a `.ps1` file in that environment.
    - Note that, the PS functions we deploy will be **constant functions** -- they cannot be modified or removed. So we don't need to worry about the function being changed on the fly.
    - `#Requires` only works when invoking the script file directly, so we don't do this deployment if `#Requires` is used in the function script and remain invoking the `.ps1` file directly in that case.
    - After turning a `.ps1` file to a PS function, script variables like `$script:xxx` used in the script will be made global variables as there is no script scope anymore. So we have to clean up global variables after every invocation.
    - Currently, we call `ResetRunspaceState()` to clean up variable table. Unfortunately, it does more than that -- it also reset the `Debugger`, create new `EventManager`, reset the `CurrentPath`. Resetting `Debugger` causes a problem to our local debugging experience -- **breakpoints set during the first invocation debugging cannot be hit in the subsequent invocations**. All we need is to clean up the global variables newly added by an invocation, so I replace `ResetRunspaceState()` with simply code that should have the similar functionality.

2. Rewrite the helper module in C#, for two reasons: (1) improve performance (2) we will need to have C# cmdlets when getting to durable function implementation anyways. This change has the following impact:
    - The helper module cannot be loaded by a version of pwsh that is lower than the PowerShell NuGet packages we reference in the worker, because of the specific version of `Newtonsoft.Json` dependency. I talked to @joeyaiello about this, and we believe it's OK to do this, as we don't think it's useful to load this helper module outside of the worker in the first place. This change won't help us to have the auto-completion UX for the helper commands in VSCode, but it's a separate issue.
    - We cannot use `Pester` to test the helper module, so I move all pester tests to xUnit.
    - We cannot do the markdown help doc check anymore (cannot load the helper module unless download the right version of `pwsh`)

Next steps
------------

- [ ] Collect feedback and concerns for scenarios and do verification
   - [x] Local debugging
      - DONE. Local debugging works as expected, and the breakpoint issue highlighted above is also resolved.
   - [x] Special variables like `$PSScriptRoot` and `$PSCommandPath`
      - DONE. They work as expected. `$PSScriptRoot` points to the script file's parent directory. `PSCommandPath` points to the script file itself.
      - xUnit test added for this scenario.
   - [x] Error should continue to report the correct script stack trace and line/column offset. (Thanks @JustinGrote!)
      - DONE. The error record works as expected. Script stack trace and line/column offset are the same as running the `.ps1` directly.
   - [ ] More ... pending on feedback
- [x] Collect the throughput data with the final changes

Throughput measurements with the final code change
---------

I measured for **20 times** with a dedicated VM (1 CPU core, 1.7 GB memory).

### Testing environment

I used a **fresh sandbox process** for every measurement with the following steps:

1. Run `net stop dwassvc` to stop the service
2. Run `net start dwassvc` to start the service
3. Wait for CPU to come down
4. Run `irm <http-trigger-uri>` for 3,4 times to pass the cold start.
5. Start throughput test using `WCAT Runner`, with this setting: **100 Users, 1 minute Duration**
    - `WCAT` first uses 30 seconds to warm up the service, then collects the **average Response Per Second (RPS)** value every 10 seconds for 1 minute

### Testing FunctionApp

An HTTP function that returns a OK response for every invocation.

```powershell
param($req, $TriggerMetadata)
Push-OutputBinding -Name res -Value ([HttpResponseContext]@{
  StatusCode = 200
  Body = "Hello"
})
```

### Measurement data

#### Overall statistics for the 20 measurements (the last column from the detailed data set table below)

```
PS> 74, 70, 89, 69, 85, 83, 76, 68, 68, 70, 73, 80, 74, 69, 85, 73, 84, 68, 76, 73 | Measure-Object -AllStats

Count             : 20
Average           : 75.35
Sum               : 1507
Maximum           : 89
Minimum           : 68
StandardDeviation : 6.68285315683275
```

According to Dormann 2013 CV-values below 0.05 (5%) indicate very high precision of the data, values above 0.2 (20%) low precision (a rule of thumb).
The CV value of my measurement data set is 0.0887 (8.87%), so the data are **relatively precise based on the rule of thumb**.

| Count | Sum | Maximum | Minimum | Mean | Std. Deviation | coefficient of variation (CV) |
| -------| ------| -----------| ------------| -------| ----------------| ------------------------------|
| 20 | 1507 | 89 | 68 | 75.35 | 6.683 | 0.0887 (Std. Dev. / Mean) |


#### Detailed measurement data set

`WCAT` collects the **average Response Per Second (RPS)** value every 10 seconds for 1 minute

| Id  | 5 sec | 15 sec | 25 sec | 35 sec | 45 sec | 55 sec | 60 sec |
| ----| ----| ----| ----| ----| ----| ----| ----|
| 1 | 77 | 77 | 81 | 82 | 78 | 74 | 74 |
| 2 | 73 | 73 | 73 | 72 | 72 | 70 | 70 |
| 3 | 45 | 87 | 88 | 88 | 87 | 88 | 89 |
| 4 | 60 | 64 | 69 | 70 | 69 | -- | 69 |
| 5 | 69 | 79 | 83 | 84 | 85 | 86 | 85 |
| 6 | 72 | 84 | 87 | 87 | 86 | 85 | 83 |
| 7 | 79 | 75 | 81 | 82 | 79 | 76 | 76 |
| 8 | 67 | 66 | 67 | 65 | 69 | 68 | 68 |
| 9 | 68 | 58 | 62 | 67 | 69 | 68 | 68 |
| 10 | 68 | 75 | 73 | 72 | 71 | 70 | 70 |
| 11 | 56 | 73 | 71 | 71 | 72 | 73 | 73 |
| 12 | 78 | 84 | 85 | 85 | 81 | 80 | 80 |
| 13 | 73 | 77 | 81 | 81 | 76 | 74 | 74 |
| 14 | 54 | 66 | 69 | 68 | 69 | 68 | 69 |
| 15 | 77 | 84 | 85 | 87 | 87 | 87 | 85 |
| 16 | 72 | 74 | 73 | 72 | 72 | 72 | 73 |
| 17 | 75 | 72 | 78 | 82 | 84 | -- | 84 |
| 18 | 59 | 61 | 66 | 66 | 69 | 68 | 68 |
| 19 | 84 | 73 | 75 | 78 | 77 | 76 | 76 |
| 20 | 54 | 71 | 71 | 73 | 73 | 72 | 73 |

### Measurement data WITHOUT changes in this PR

Before this PR, the average RPS is 22.3. This PR **achieves over 3 time improvement (75.35 / 22.3 = 3.37)**

| Time | Test | Runtime | RPS | Total requests | Average response time |
| ---| ---| ---| ---| ---| ---|
|2019/05/13 14:34:26 | PS Ping | 2.0.12455.0 | 20 (Threshold:20) | 2468 | 4941.86|
|2019/05/12 14:35:25 | PS Ping | 2.0.12455.0 | 21 (Threshold:20) | 2543 | 4792.75|
|2019/05/11 14:33:38 | PS Ping | 2.0.12455.0 | 23 (Threshold:20) | 2878 | 4219.34|
|2019/05/10 14:33:26 | PS Ping | 2.0.12453.0 | 24 (Threshold:20) | 2889 | 4201.02|
|2019/05/09 14:34:31 | PS Ping | 2.0.12451.0 | 21 (Threshold:20) | 2530 | 4797.76|
|2019/05/08 14:33:00 | PS Ping | 2.0.12451.0 | 22 (Threshold:20) | 2756 | 4405.71|
|2019/05/07 14:36:42 | PS Ping | 2.0.12451.0 | 23 (Threshold:20) | 2819 | 4305.09|
|2019/05/06 14:33:37 | PS Ping | 2.0.12451.0 | 24 (Threshold:20) | 2947 | 4119.65|
|2019/05/05 14:34:22 | PS Ping | 2.0.12451.0 | 22 (Threshold:20) | 2681 | 4529.3|
|2019/05/04 14:34:27 | PS Ping | 2.0.12451.0 | 23 (Threshold:20) | 2865 | 4238.09|
|2019/05/03 14:34:24 | PS Ping | 2.0.12449.0 | 22 (Threshold:20) | 2758 | 4399.92|
|2019/05/02 14:33:10 | PS Ping | 2.0.12447.0 | 23 (Threshold:20) | 2841 | 4273.55|
|2019/05/01 18:14:40 | PS Ping | 2.0.12447.0 | 22 (Threshold:20) | 2675 | 4541.99|

### Future work on throughput

@pragnagopa has been working on analyzing how to boost workers' throughput in the gRPC layer, basically around how workers should handle incoming/outgoing gRPC messages more efficiently.

This PR focused on fixing the bottleneck in the PowerShell specific logic, and the future work should shift focus to the gRPC message processing in the PowerShell worker. Open #224 to track the future work.

Also, all the measurements and analysis I did for this PR was using the default `PSWorkerInProcConcurrencyUpperBound` value, which is 1, meaning that invocations are not processed concurrently. This default behavior makes sense for the consumption plan, given the limited processing power of the VM used for consumption plan. However, for the premium plan, the VM has much better processing power, and thus the throughput should be analyzed again in that environment with the concurrency support enabled.